[Bronze V] Title: A/B, Time: 176 ms, Memory: 17796 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.java
@@ -4,8 +4,8 @@ public class Main {
     
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextInt();
+        int a = sc.nextInt();
         int b = sc.nextInt();
-        System.out.println(a/b);
+        System.out.println((double)a/b);
     }
 }
